+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.0.1
+
+- (Feat) remove gradle caches using (--gradle-cache || -c)
+
+## 2.0.0
+
+- arguments rework
+
 ## 1.0.0
 
 - Initial version.